Emoticon 클래스를 만들어 따로 분리
